requirements.txt
python-telegram-bot==20.7
reportlab==4.0.8
Pillow==10.2.0
python-dotenv==1.0.0
openai-whisper==20231117
numpy==1.26.3
torch==2.1.2
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
python-multipart==0.0.6
